% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{boj_frequencies}
\alias{boj_frequencies}
\title{Available Frequency Codes}
\format{
A named character vector.
}
\usage{
boj_frequencies
}
\description{
A named character vector mapping frequency codes to their descriptions.
}
\keyword{datasets}
