% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boj_request}
\alias{boj_request}
\title{Make a BOJ API request}
\usage{
boj_request(endpoint, params, request_interval = 1, timeout = 30)
}
\arguments{
\item{endpoint}{API endpoint name (e.g. "getDataCode")}

\item{params}{Named list of query parameters}

\item{request_interval}{Seconds to wait between requests}

\item{timeout}{Request timeout in seconds}
}
\value{
Parsed JSON response as a list
}
\description{
Make a BOJ API request
}
\keyword{internal}
