% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bojstat-package.R
\docType{package}
\name{bojstat-package}
\alias{bojstat}
\alias{bojstat-package}
\title{bojstat: R Interface to the Bank of Japan Time-Series Data Search API}
\description{
Provides functions to access the BOJ Time-Series Data Search API,
launched on February 18, 2026. Retrieve over 200,000 time-series
statistical data including interest rates, exchange rates, money stock,
Tankan survey results, price indices, and more.
}
\note{
This package is not officially affiliated with or endorsed by
the Bank of Japan. Please avoid high-frequency access to the API,
as the BOJ may block connections if requests are too frequent.
Use the \code{request_interval} parameter (default: 1 second) appropriately.
}
\section{Main functions}{

\itemize{
\item \code{\link[=boj_get_data]{boj_get_data()}} -- Code API: retrieve data by series codes
\item \code{\link[=boj_get_data_all]{boj_get_data_all()}} -- Code API with automatic pagination
\item \code{\link[=boj_get_layer]{boj_get_layer()}} -- Layer API: retrieve data by hierarchical structure
\item \code{\link[=boj_get_metadata]{boj_get_metadata()}} -- Metadata API: retrieve series metadata
\item \code{\link[=boj_search_series]{boj_search_series()}} -- Search series by keyword
\item \code{\link[=boj_list_databases]{boj_list_databases()}} -- List all available databases
}
}

\section{Reference data}{

\itemize{
\item \link{boj_db_names} -- Named vector of all database codes and descriptions
\item \link{boj_frequencies} -- Named vector of frequency codes
}
}

\references{
\itemize{
\item API announcement: \url{https://www.boj.or.jp/statistics/outline/notice_2026/not260218a.htm}
\item API manual: \url{https://www.stat-search.boj.or.jp/info/api_manual.pdf}
\item BOJ Time-Series Data Search: \url{https://www.stat-search.boj.or.jp}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kigasudayooo/bojstat}
  \item Report bugs at \url{https://github.com/kigasudayooo/bojstat/issues}
}

}
\author{
\strong{Maintainer}: Kazuki Hosoda \email{kazu.1215.sagittarius@gmail.com}

}
